backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
stripe>=10.0.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
